Re-implement useDisplayPayee to use context to minimize SQL queries
