+++ packages/desktop-client/src/components/mobile/transactions/TransactionListWithBalances.tsx
@@ -16,6 +16,7 @@ import {
   CellValue,
   CellValueText,
 } from '@desktop-client/components/spreadsheet/CellValue';
+import { DisplayPayeeContextProvider } from '@desktop-client/hooks/useDisplayPayee';
 import {
   SelectedProvider,
   useSelected,
@@ -109,8 +110,8 @@ export function TransactionListWithBalances({
   const selectedInst = useSelected('transactions', [...transactions], []);
 
   return (
-    <SelectedProvider instance={selectedInst}>
-      <>
+    <DisplayPayeeContextProvider transactions={transactions}>
+      <SelectedProvider instance={selectedInst}>
         <View
           style={{
             flexShrink: 0,
@@ -151,8 +152,8 @@ export function TransactionListWithBalances({
             showMakeTransfer={showMakeTransfer}
           />
         </PullToRefresh>
-      </>
-    </SelectedProvider>
+      </SelectedProvider>
+    </DisplayPayeeContextProvider>
   );
 }
