+++ upcoming-release-notes/5795.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [joel-jeremy]
+---
+
+Re-implement useDisplayPayee to use context to minimize SQL queries
